requirements.txt
python-telegram-bot
transformers
torch
torchvision
torchaudio
pymongo
sounddevice
numpy
librosa
requests